Changed cross-compilation mode.
Instead of compiling against a bootstrap classpath (i.e. a downloadable
rt.jar file) using the -source and -target compiler switches, it uses
the newer -release option, which does not require a version of rt.jar,
and supports cross-compiling to versions of Java all the way down to
version 6.

This option is only available for compilers of Java 9 and higher.
Consequently, cross-compiling will no longer work on systems with Java 8
and lower, *but* one can still compile JAR files that will be compatible
with Java 6.